requirements.txt
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distlib==0.3.9
executing==2.1.0
filelock==3.16.1
frida==16.5.9
frida-tools==13.6.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
Sastrawi==1.0.1
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
urllib3==2.2.3
virtualenv==20.28.0
wcwidth==0.2.13
websockets==13.1
